@@ -5,8 +5,8 @@ def safeExtGet(prop, fallback) {
 }
 
 android {
-    compileSdkVersion safeExtGet('compileSdkVersion', 27)
-    buildToolsVersion safeExtGet('buildToolsVersion', '27.0.3')
+    compileSdkVersion safeExtGet('compileSdkVersion', 28)
+    buildToolsVersion safeExtGet('buildToolsVersion', '28.0.3')
 
     compileOptions {
         targetCompatibility JavaVersion.VERSION_1_8
@@ -15,7 +15,7 @@ android {
 
     defaultConfig {
         minSdkVersion safeExtGet('minSdkVersion', 16)
-        targetSdkVersion safeExtGet('targetSdkVersion', 27)
+        targetSdkVersion safeExtGet('targetSdkVersion', 28)
         versionCode 1
         versionName "1.0"
     }
@@ -27,19 +27,25 @@ android {
 }
 
 dependencies {
-    compileOnly "com.facebook.react:react-native:${safeExtGet('reactNativeVersion', '+')}"
-    implementation('com.google.android.exoplayer:exoplayer:2.9.3') {
+    implementation "com.facebook.react:react-native:${safeExtGet('reactNativeVersion', '+')}"
+    implementation('com.google.android.exoplayer:exoplayer:2.11.8') {
         exclude group: 'com.android.support'
     }
 
     // All support libs must use the same version
-    implementation "com.android.support:support-annotations:${safeExtGet('supportLibVersion', '+')}"
-    implementation "androidx.appcompat:appcompat:1.1.0"
-    implementation "com.android.support:support-media-compat:${safeExtGet('supportLibVersion', '+')}"
+    // DRM replaced the dependencies bellow
+    // implementation "com.android.support:support-annotations:${safeExtGet('supportLibVersion', '+')}"
+    // implementation "androidx.appcompat:appcompat:1.1.0"
+    // implementation "com.android.support:support-media-compat:${safeExtGet('supportLibVersion', '+')}"
+    implementation "androidx.annotation:annotation:1.1.0"
+    implementation "androidx.core:core:1.1.0"
+    implementation "androidx.media:media:1.1.0"
 
-    implementation('com.google.android.exoplayer:extension-okhttp:2.9.3') {
+    implementation('com.google.android.exoplayer:extension-okhttp:2.11.4') {
         exclude group: 'com.squareup.okhttp3', module: 'okhttp'
     }
-    implementation 'com.squareup.okhttp3:okhttp:3.12.1'
+    implementation 'com.squareup.okhttp3:okhttp:3.14.3'
 
+    // MediaSession extension
+    implementation 'com.google.android.exoplayer:extension-mediasession:2.8.1'
 }
