@@ -2,15 +2,18 @@ package com.brentvatne.exoplayer;
 
 import android.annotation.TargetApi;
 import android.content.Context;
-import android.support.v4.content.ContextCompat;
+import androidx.core.content.ContextCompat;
 import android.util.AttributeSet;
 import android.util.Log;
+import android.util.TypedValue;
 import android.view.Gravity;
 import android.view.SurfaceView;
 import android.view.TextureView;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
+
+import android.view.accessibility.CaptioningManager;
 
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.ExoPlaybackException;
@@ -21,13 +24,22 @@ import com.google.android.exoplayer2.Timeline;
 import com.google.android.exoplayer2.source.TrackGroupArray;
 import com.google.android.exoplayer2.text.Cue;
 import com.google.android.exoplayer2.text.TextRenderer;
+import com.google.android.exoplayer2.text.TextOutput;
 import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
 import com.google.android.exoplayer2.ui.SubtitleView;
+import com.google.android.exoplayer2.util.Util;
 
 import java.util.List;
 
 @TargetApi(16)
 public final class ExoPlayerView extends FrameLayout {
+
+    /**
+     * The default fractional text size.
+     *
+     * @see #setFractionalTextSize(float, boolean)
+     */
+    public static final float DEFAULT_TEXT_SIZE_FRACTION = 0.0533f;
 
     private View surfaceView;
     private final View shutterView;
@@ -37,6 +49,7 @@ public final class ExoPlayerView extends FrameLayout {
     private SimpleExoPlayer player;
     private Context context;
     private ViewGroup.LayoutParams layoutParams;
+    private float textTrackSizeScale;
 
     private boolean useTextureView = true;
     private boolean hideShutterView = false;
@@ -75,6 +88,7 @@ public final class ExoPlayerView extends FrameLayout {
         subtitleLayout.setLayoutParams(layoutParams);
         subtitleLayout.setUserDefaultStyle();
         subtitleLayout.setUserDefaultTextSize();
+        configureSubtitleFontSize();
 
         updateSurfaceView();
 
@@ -82,6 +96,33 @@ public final class ExoPlayerView extends FrameLayout {
         layout.addView(subtitleLayout, 2, layoutParams);
 
         addViewInLayout(layout, 0, aspectRatioParams);
+    }
+
+    public void setTextTrackSizeScale(float textTrackSizeScale) {
+        this.textTrackSizeScale = textTrackSizeScale;
+
+        configureSubtitleFontSize();
+    }
+
+    public void configureSubtitleFontSize() {
+        if (subtitleLayout == null) {
+            return;
+        }
+
+        if (textTrackSizeScale <= 0) {
+            return;
+        }
+
+        float userFontScale = Util.SDK_INT >= 19 && !subtitleLayout.isInEditMode() ? getUserCaptionFontScaleV19() : 1f;
+
+        subtitleLayout.setFractionalTextSize(DEFAULT_TEXT_SIZE_FRACTION * userFontScale * textTrackSizeScale);
+    }
+
+    @TargetApi(19)
+    private float getUserCaptionFontScaleV19() {
+        CaptioningManager captioningManager =
+                (CaptioningManager) getContext().getSystemService(Context.CAPTIONING_SERVICE);
+        return captioningManager.getFontScale();
     }
 
     private void setVideoView() {
@@ -200,7 +241,7 @@ public final class ExoPlayerView extends FrameLayout {
     }
 
     private final class ComponentListener implements SimpleExoPlayer.VideoListener,
-            TextRenderer.Output, ExoPlayer.EventListener {
+            TextOutput, ExoPlayer.EventListener {
 
         // TextRenderer.Output implementation
 
@@ -213,13 +254,8 @@ public final class ExoPlayerView extends FrameLayout {
 
         @Override
         public void onVideoSizeChanged(int width, int height, int unappliedRotationDegrees, float pixelWidthHeightRatio) {
-            boolean isInitialRatio = layout.getAspectRatio() == 0;
             layout.setAspectRatio(height == 0 ? 1 : (width * pixelWidthHeightRatio) / height);
-
-            // React native workaround for measuring and layout on initial load.
-            if (isInitialRatio) {
-                post(measureAndLayout);
-            }
+            post(measureAndLayout);
         }
 
         @Override
