@@ -1,5 +1,5 @@
 export default {
   SRT: 'application/x-subrip',
   TTML: 'application/ttml+xml',
-  VTT: 'text/vtt'
+  VTT: 'text/vtt',
 };
