@@ -1,5 +1,7 @@
 package com.brentvatne.react;
 
+import com.brentvatne.exoplayer.DefaultReactExoplayerConfig;
+import com.brentvatne.exoplayer.ReactExoplayerConfig;
 import com.brentvatne.exoplayer.ReactExoplayerViewManager;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.JavaScriptModule;
@@ -12,19 +14,31 @@ import java.util.List;
 
 public class ReactVideoPackage implements ReactPackage {
 
+    private ReactExoplayerConfig config;
+
+    public ReactVideoPackage() {
+    }
+
+    public ReactVideoPackage(ReactExoplayerConfig config) {
+        this.config = config;
+    }
+
     @Override
     public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
         return Collections.emptyList();
     }
 
-    // Deprecated RN 0.47	
-    public List<Class<? extends JavaScriptModule>> createJSModules() {	
-        return Collections.emptyList();	
-    }	
+    // Deprecated RN 0.47
+    public List<Class<? extends JavaScriptModule>> createJSModules() {
+        return Collections.emptyList();
+    }
 
 
     @Override
     public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
-        return Collections.<ViewManager>singletonList(new ReactExoplayerViewManager());
+        if (config == null) {
+            config = new DefaultReactExoplayerConfig(reactContext);
+        }
+        return Collections.singletonList(new ReactExoplayerViewManager(config));
     }
 }
